added window size class state to application for responsive design. fixed weird bug involving float/double values and localisation when used by stringResources

+++ app/src/main/java/no/hiof/friluftslivcompanionapp/ui/components/items/TripItem.kt
@@ -59,7 +59,7 @@ fun TripItem(trip: Hike) {
             Column(
                 modifier = Modifier.align(Alignment.BottomStart)
             ) {
-                val roundedDistance = String.format("%.1f", trip.distanceKm).toFloat()
+                val roundedDistance = String.format("%.1f", trip.distanceKm)
                 Text(
                     text = "Distance: $roundedDistance km",
                     color = MaterialTheme.colorScheme.surface,
